changed name from customOperations to customFunctions and wrote a couple more tests
