+++ src/engine/executors/aggregate-executor.ts
@@ -48,13 +48,13 @@ export default class AggregateExecutor extends Executor {
    * @param options - Execution options
    * @return An iterator which evaluate SPARQL aggregations
    */
-  buildIterator (source: Observable<Bindings>, query: Algebra.RootNode, context: ExecutionContext, customOperators: any): Observable<Bindings> {
+  buildIterator (source: Observable<Bindings>, query: Algebra.RootNode, context: ExecutionContext, customFunctions?: any): Observable<Bindings> {
     if ('group' in query) {
       // first, group bindings using the GROUP BY clause
-      let iterator = this._executeGroupBy(source, query.group || [], context, customOperators)
+      let iterator = this._executeGroupBy(source, query.group || [], context, customFunctions)
       // next, apply the optional HAVING clause to filter groups
       if ('having' in query) {
-        iterator = this._executeHaving(iterator, query.having || [], context, customOperators)
+        iterator = this._executeHaving(iterator, query.having || [], context, customFunctions)
       }
       return iterator
     }
@@ -68,7 +68,7 @@ export default class AggregateExecutor extends Executor {
    * @param  options - Execution options
    * @return An iterator which evaluate a GROUP BY clause
    */
-  _executeGroupBy (source: Observable<Bindings>, groupby: Algebra.Aggregation[], context: ExecutionContext, customOperators: any): Observable<Bindings> {
+  _executeGroupBy (source: Observable<Bindings>, groupby: Algebra.Aggregation[], context: ExecutionContext, customFunctions?: any): Observable<Bindings> {
     let iterator = source
     // extract GROUP By variables & rewrite SPARQL expressions into BIND clauses
     const variables: string[] = []
@@ -77,7 +77,7 @@ export default class AggregateExecutor extends Executor {
         variables.push(g.expression)
       } else {
         variables.push(g.variable)
-        iterator = iterator.pipe(bind(g.variable, g.expression, customOperators))
+        iterator = iterator.pipe(bind(g.variable, g.expression, customFunctions))
       }
     })
     return groupBy(iterator, variables)
@@ -90,11 +90,11 @@ export default class AggregateExecutor extends Executor {
    * @param  options - Execution options
    * @return An iterator which evaluate a HAVING clause
    */
-  _executeHaving (source: Observable<Bindings>, having: Algebra.Expression[], context: ExecutionContext, customOperators: any): Observable<Bindings> {
+  _executeHaving (source: Observable<Bindings>, having: Algebra.Expression[], context: ExecutionContext, customFunctions?: any): Observable<Bindings> {
     // thanks to the flexibility of SPARQL expressions,
     // we can rewrite a HAVING clause in a set of FILTER clauses!
     return having.reduce((iter, expression) => {
-      return iter.pipe(filter(expression, customOperators))
+      return iter.pipe(filter(expression, customFunctions))
     }, source)
   }
 }
